requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
pandas==2.2.2
openpyxl==3.1.2
PyPDF2==3.0.0
pdfplumber==0.7.6
python-docx==0.8.11
stripe==6.0.0
gunicorn==21.2.0
